refactor: вынесен функционал определения компонентов
добавлена функция создания компонентов на основе шаблонизатора
Handlebars
